@@ -4,34 +4,31 @@ import mu.KotlinLogging
 import org.springframework.beans.factory.DisposableBean
 import org.springframework.boot.ApplicationArguments
 import org.springframework.boot.ApplicationRunner
-import org.telegram.telegrambots.meta.TelegramBotsApi
-import org.telegram.telegrambots.meta.generics.BotSession
-import org.telegram.telegrambots.meta.generics.LongPollingBot
-import org.telegram.telegrambots.updatesreceivers.DefaultBotSession
+import org.telegram.telegrambots.longpolling.TelegramBotsLongPollingApplication
+import org.telegram.telegrambots.longpolling.util.LongPollingSingleThreadUpdateConsumer
 
 private val log = KotlinLogging.logger {}
 
 open class TgBotRunner(
-    private val tgBot: LongPollingBot,
+    private val tgBot: LongPollingSingleThreadUpdateConsumer,
     private val props: TgBotProperties
 ) : ApplicationRunner, DisposableBean {
 
-    private val telegramBotApi = TelegramBotsApi(DefaultBotSession::class.java)
-    private var session: BotSession? = null
+    private val botsApplication = TelegramBotsLongPollingApplication()
 
     override fun run(args: ApplicationArguments?) {
         if (props.runnable) {
-            this.session = telegramBotApi.registerBot(tgBot)
-            log.info { "Bot '${tgBot.botUsername}' started" }
+            botsApplication.registerBot(props.token, tgBot)
+            log.info { "Bot '${props.username}' started" }
         } else {
-            log.info { "Bot '${tgBot.botUsername}' is not runnable" }
+            log.info { "Bot '${props.username}' is not runnable" }
         }
     }
 
     override fun destroy() {
         if (props.gracefulShutdown) {
-            this.session?.stop()
-            log.info { "Bot '${tgBot.botUsername}' stopped" }
+            this.botsApplication.close()
+            log.info { "Bot '${props.username}' stopped" }
         }
     }
 }
