@@ -1,20 +1,14 @@
 package ru.v1as.tg.starter
 
-import org.telegram.telegrambots.bots.DefaultBotOptions
-import org.telegram.telegrambots.bots.TelegramLongPollingBot
+import org.telegram.telegrambots.longpolling.util.LongPollingSingleThreadUpdateConsumer
 import org.telegram.telegrambots.meta.api.objects.Update
 import ru.v1as.tg.starter.update.UpdateProcessor
 
 open class TgLongPollingBot(
-    options: DefaultBotOptions,
-    private val props: TgBotProperties,
     private val updateProcessor: UpdateProcessor
-) :
-    TelegramLongPollingBot(options, props.token) {
+) : LongPollingSingleThreadUpdateConsumer {
 
-    override fun getBotUsername() = props.username
-
-    override fun onUpdateReceived(update: Update?) {
+    override fun consume(update: Update?) {
         if (update != null) {
             updateProcessor.process(update)
         }
