@@ -1,6 +1,6 @@
 package ru.v1as.tg.starter
 
-import org.telegram.telegrambots.meta.api.objects.Message
+import org.telegram.telegrambots.meta.api.objects.message.Message
 import ru.v1as.tg.starter.model.TgChat
 import ru.v1as.tg.starter.model.TgUser
 import java.util.concurrent.CompletableFuture
