@@ -2,6 +2,7 @@ package ru.v1as.tg.starter.keyboard
 
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton
+import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardRow
 import ru.v1as.tg.starter.TgBotProperties
 
 class StartCommandLinkFactory(val tgBot: TgBotProperties) {
@@ -14,5 +15,5 @@ class StartCommandLinkFactory(val tgBot: TgBotProperties) {
     }
 
     fun buildKeyboard(text: String, vararg arguments: Any) =
-        InlineKeyboardMarkup(listOf(listOf(buildButton(text, *arguments))))
+        InlineKeyboardMarkup(listOf(InlineKeyboardRow(buildButton(text, *arguments))))
 }
