@@ -1,7 +1,7 @@
 package ru.v1as.tg.starter.update.command
 
-import org.telegram.telegrambots.meta.api.objects.Message
 import org.telegram.telegrambots.meta.api.objects.Update
+import org.telegram.telegrambots.meta.api.objects.message.Message
 import ru.v1as.tg.starter.model.base.TgChatWrapper
 import ru.v1as.tg.starter.model.base.TgUserWrapper
 import ru.v1as.tg.starter.update.UpdateRequest
